+++ http_misc/logger.py
@@ -1,3 +1,6 @@
+"""
+Утилиты логирования
+"""
 import logging
 
 _logger = logging.getLogger(__name__)